3 Beginner AI Projects/project1-BasicAIAgent/pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.1",
]
